Expose handshake_timeout option in listeners
Ranch had a default value of 5 seconds for the handshake timeout. This sometimes
caused constrained devices to fail the connection since they couldn't complete
the SSL handshake in that amount of time.

This commit allows configuring the handshake timeout to arbitrary values to
support such use cases.

This also includes a small internal refactoring, renaming the old transport_opts to
socket_opts, since that's what they actually are, and introducing new base
transport_opts that depend on the ranch transport.

Fix #1670

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>